Blogilistan laajennus, step 10

+++ osa4/blogilista/controllers/blogs.js
@@ -67,6 +67,31 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  // Get user
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  // Get blog post
+  const blog = await Blog.findById(request.params.id)
+
+  // console.log(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({
+      error: 'Blog post not found '
+    })
+  }
+  // Check that only user who created blog post can delete it
+  if (blog.user._id.toString() !== user.id.valueOf()) {
+    return response.status(401).json({
+      error: 'user not a post creator'
+    })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
